refactor(protocols): collapse protocol split onto aragora.protocols (VAL-P4A-009)
Move backend Protocol definitions into aragora/protocols/backend_protocols.py
and re-export the full collapsed surface from aragora.protocols. Convert
aragora/core_protocols.py and aragora/type_protocols.py into two-sided
DeprecationWarning shims: a plain import succeeds, an import under
-W error::DeprecationWarning fails, and the warning text carries the
core_protocols / type_protocols tokens. Migrate Tier-1 internal callers onto
aragora.protocols; defer the callers that sit on pre-existing mypy union-attr
type-debt (the shim keeps their old import path working).
