+++ aragora/debate/arena_primary_configs.py
@@ -11,7 +11,7 @@ from aragora.debate.protocol import CircuitBreaker
 if TYPE_CHECKING:
     from aragora.debate.protocol import DebateProtocol
 from aragora.spectate.stream import SpectatorStream
-from aragora.type_protocols import (
+from aragora.protocols import (
     BroadcastPipelineProtocol,
     CalibrationTrackerProtocol,
     ConsensusMemoryProtocol,
