+++ aragora/debate/phases/feedback_calibration.py
@@ -13,7 +13,7 @@ from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
     from aragora.debate.context import DebateContext
-    from aragora.type_protocols import CalibrationTrackerProtocol, EventEmitterProtocol
+    from aragora.protocols import CalibrationTrackerProtocol, EventEmitterProtocol
 
 logger = logging.getLogger(__name__)
